+++ js/validation.js
@@ -1,0 +1,3 @@
+export function checkMaxLength(collection, maxLength) {
+  return collection.length !== undefined && collection.length <= maxLength;
+}
